pyproject.toml
[tool.poetry]
name = "work-base-date"
version = "0.1.0"
description = ""
authors = ["Daria"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.11"
requests = "^2.32.5"
google-api-python-client = "^2.185.0"
coverage = "^7.11.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.18.2"
black = "^25.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
